Some bug fixes to control LEDs when Pad ID option is set
In this version LED will be on if Ds3LEDsFunction is Pad ID and option
"Flash when charging or dieing" is not checked. It will be flashing if
option is checked and battery is charging with USB cable. When battery
is charged it will be on. And it will be flashing when connected with
BTH and battery state is Low. So it will be functioning as in old
version before my pull request.